Don't overload Perlbal headers to get paths
Using Perlbal headers to communicate the paths of a file wasn't a very
smart idea of mine, since Perlbal will interpret them if it happens to
be in front of the mogrest server. NOTE we still want to avoid mogrest
to server the files itself if we can, so mogrest in itself supports
x-reproxy protocol.